requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
requests==2.32.3
urllib3==2.2.3
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
